Wurm Rest Client v0.1.0 WurmRestClient

Parent module for the client for the Wurm REST API as implemented by wurmrest. This module provides common functions for retrieving from the API and transformation of the response into basic enumerated result types.

Functions

fetch(url)

Fetch a URL and return the response, transformed to a result type with body.

handle_response(url, arg)

Given a response to a request, returns a tuple depending on the response. If the API call was successfully completed, will return a response in the following form and its body as parsed from JSON:

{ result_type, response_body}

A call may be successfully completed even if the result of that call was not success, for example if the player was not found. If the call was unsuccessful, returns a tuple in the following form: { :error_connect, reason } Where reason is an atom representing the reason for the call failure.

Example

iex> handle_response(“http://localhost:8080/players/Bob/money”, {:ok, %{status_code: 200, body: “{“balance”: 260 }}) %{:ok, %{balance: 260}} iex> handle_response(“http://localhost:8080/players/Anne/money”, {:error, %HTTPoison.Error{reason: :timeout}}) %{:error_connect, :timeout}