Wurm Rest Client v0.1.0 WurmRestClient.Player

Retrieves player-related information.

Functions

fetch_money(player)

Given a player, fetches that player’s current bank balance.

handle_response(arg)

Given a response to a Player request, returns a tuple depending on the response. If the API call was successfully completed, will return a response in the following form and its body as parsed from JSON:

{ result_type, response_body}

A call may be successfully completed even if the result of that call was not success, for example if the player was not found. If the call was unsuccessful, returns a tuple in the following form: { :error_connect, reason } Where reason is an atom representing the reason for the call failure.

Example

iex> handle_response({:ok, %{status_code: 200, body: “{“balance”: 260 }}) %{:ok, %{balance: 260}} iex> handle_response({:error, %HTTPoison.Error{reason: :timeout}}) %{:error_connect, :timeout}

player_money_url(player)

Given a player, returns the URL that corresponds to the location of the money resource for that player

player_url(player)

Given a player, returns the URL that corresponds to that player